requirements.txt
reflex
mysqlclient
mysql-connector-python